// sdks/node-ts/src/apache_beam/proto/standard_window_fns.ts
export interface FixedWindowsPayload {
    /**
     * (Required) Represents the size of the window.
     *
     * @generated from protobuf field: google.protobuf.Duration size = 1;
     */
    size?: Duration;
    /**
     * (Required) Represents the timestamp of when the first window begins.
     * Window N will start at offset + N * size.
     *
     * @generated from protobuf field: google.protobuf.Timestamp offset = 2;
     */
    offset?: Timestamp;
}
// ...
export interface SessionWindowsPayload {
    /**
     * (Required) Minimum duration of gaps between sessions.
     *
     * @generated from protobuf field: google.protobuf.Duration gap_size = 1;
     */
    gapSize?: Duration;
}
// ...
class FixedWindowsPayload$Type extends MessageType<FixedWindowsPayload> {
    constructor() {
        super("org.apache.beam.model.pipeline.v1.FixedWindowsPayload", [
            { no: 1, name: "size", kind: "message", T: () => Duration },
            { no: 2, name: "offset", kind: "message", T: () => Timestamp }
        ]);
    }
    create(value?: PartialMessage<FixedWindowsPayload>): FixedWindowsPayload {
        const message = {};
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<FixedWindowsPayload>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: FixedWindowsPayload): FixedWindowsPayload {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* google.protobuf.Duration size */ 1:
                    message.size = Duration.internalBinaryRead(reader, reader.uint32(), options, message.size);
                    break;
                case /* google.protobuf.Timestamp offset */ 2:
                    message.offset = Timestamp.internalBinaryRead(reader, reader.uint32(), options, message.offset);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: FixedWindowsPayload, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* google.protobuf.Duration size = 1; */
        if (message.size)
            Duration.internalBinaryWrite(message.size, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* google.protobuf.Timestamp offset = 2; */
        if (message.offset)
            Timestamp.internalBinaryWrite(message.offset, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message org.apache.beam.model.pipeline.v1.FixedWindowsPayload
 */
export const FixedWindowsPayload = new FixedWindowsPayload$Type();
// ...
class SessionWindowsPayload$Type extends MessageType<SessionWindowsPayload> {
    constructor() {
        super("org.apache.beam.model.pipeline.v1.SessionWindowsPayload", [
            { no: 1, name: "gap_size", kind: "message", T: () => Duration }
        ]);
    }
    create(value?: PartialMessage<SessionWindowsPayload>): SessionWindowsPayload {
        const message = {};
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<SessionWindowsPayload>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SessionWindowsPayload): SessionWindowsPayload {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* google.protobuf.Duration gap_size */ 1:
                    message.gapSize = Duration.internalBinaryRead(reader, reader.uint32(), options, message.gapSize);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SessionWindowsPayload, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* google.protobuf.Duration gap_size = 1; */
        if (message.gapSize)
            Duration.internalBinaryWrite(message.gapSize, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message org.apache.beam.model.pipeline.v1.SessionWindowsPayload
 */
export const SessionWindowsPayload = new SessionWindowsPayload$Type();